# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "examples": [
              "VALIDATION_ERROR"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetailOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "examples": [
              "One or more fields are invalid."
            ],
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "examples": [
              "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
            ],
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetailOut": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetailOut",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PageOut_UserOut_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "examples": [
              1
            ],
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "examples": [
              20
            ],
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "examples": [
              134
            ],
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "title": "PageOut[UserOut]",
        "type": "object"
      },
      "Preferences": {
        "additionalProperties": false,
        "properties": {
          "theme": {
            "$ref": "#/components/schemas/Theme",
            "default": "system"
          }
        },
        "title": "Preferences",
        "type": "object"
      },
      "Priority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "Role": {
        "enum": [
          "developer",
          "lead"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "SummaryOut": {
        "additionalProperties": false,
        "properties": {
          "activeProjects": {
            "minimum": 0.0,
            "title": "Activeprojects",
            "type": "integer"
          },
          "completionRate": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Completionrate",
            "type": "integer"
          },
          "openTasks": {
            "minimum": 0.0,
            "title": "Opentasks",
            "type": "integer"
          },
          "overdueTasks": {
            "minimum": 0.0,
            "title": "Overduetasks",
            "type": "integer"
          },
          "upcomingDeadlines": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "title": "Upcomingdeadlines",
            "type": "array"
          }
        },
        "required": [
          "activeProjects",
          "openTasks",
          "overdueTasks",
          "completionRate",
          "upcomingDeadlines"
        ],
        "title": "SummaryOut",
        "type": "object"
      },
...
      "TaskOut": {
        "additionalProperties": false,
        "properties": {
          "assigneeId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigneeid"
          },
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "description": "UTC, ISO 8601, ends in Z.",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dueDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "description": "UTC, ISO 8601, ends in Z.",
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "projectId",
          "title",
          "description",
          "status",
          "priority",
          "dueDate",
          "assigneeId",
          "completedAt"
        ],
        "title": "TaskOut",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "todo",
          "in_progress",
          "in_review",
          "done"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
      "Theme": {
        "enum": [
          "light",
          "dark",
          "system"
        ],
        "title": "Theme",
        "type": "string"
      },
...
      "UserCreate": {
        "additionalProperties": false,
        "properties": {
          "avatarUrl": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl"
          },
          "email": {
            "examples": [
              "ada@example.com"
            ],
            "format": "email",
            "maxLength": 254,
            "title": "Email",
            "type": "string"
          },
          "name": {
            "examples": [
              "Ada Lovelace"
            ],
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "description": "12 to 128 characters, not equal to the email.",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserOut": {
        "additionalProperties": false,
        "properties": {
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl"
          },
          "createdAt": {
            "description": "UTC, ISO 8601, ends in Z.",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "updatedAt": {
            "description": "UTC, ISO 8601, ends in Z.",
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "name",
          "email",
          "role",
          "avatarUrl",
          "preferences"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A token from `POST /api/v1/auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nThe backend for the DevDash dashboard: users, projects, tasks, activity and a dashboard summary.\n\n**Try it in five steps**\n\n1. `POST /api/v1/users` to register (or use a seeded account in development).\n2. `POST /api/v1/auth/login` to get a token.\n3. Click **Authorize** above and paste the `accessToken`.\n4. `POST /api/v1/projects`, then `POST /api/v1/tasks`.\n5. `PATCH /api/v1/tasks/{taskId}/status` to move a task through its workflow.\n\nEvery error has the same shape: `{\"error\": {\"code\", \"message\", \"details\"?, \"requestId\"}}`.\nState is kept in memory and resets on restart (Task 3 adds the database).\n",
    "title": "DevDash API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard/summary": {
      "get": {
        "description": "Active projects, open tasks, overdue tasks, completion rate, and the open tasks due in the next 7 days, soonest first.",
        "operationId": "summary",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "summary": "UNAUTHENTICATED",
                    "value": {
                      "error": {
                        "code": "UNAUTHENTICATED",
                        "message": "The access token is missing, invalid or expired.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "summary": "INTERNAL_ERROR",
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "Something went wrong on our side.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dashboard summary",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Paginated. Filter by `role`; search `q` matches name and email, case-insensitively.",
        "operationId": "list_users",
        "parameters": [
          {
            "description": "Page number, from 1.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number, from 1.",
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page, at most 100.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page, at most 100.",
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "description": "`field` ascending or `-field` descending.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`field` ascending or `-field` descending.",
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_UserOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "summary": "UNAUTHENTICATED",
                    "value": {
                      "error": {
                        "code": "UNAUTHENTICATED",
                        "message": "The access token is missing, invalid or expired.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "details": [
                          {
                            "field": "title",
                            "message": "Must be between 1 and 120 characters."
                          }
                        ],
                        "message": "One or more fields are invalid.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "summary": "INTERNAL_ERROR",
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "Something went wrong on our side.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Create an account. No token is needed. The role is always `developer`; only a lead can change it later. The password must be 12 to 128 characters and not equal to the email. Returns `201` with a `Location` header. Limited to 5 attempts per minute per client and email.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "MALFORMED_REQUEST": {
                    "summary": "MALFORMED_REQUEST",
                    "value": {
                      "error": {
                        "code": "MALFORMED_REQUEST",
                        "message": "The request body is not valid JSON.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed request"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "EMAIL_ALREADY_EXISTS": {
                    "summary": "EMAIL_ALREADY_EXISTS",
                    "value": {
                      "error": {
                        "code": "EMAIL_ALREADY_EXISTS",
                        "message": "An account with this email already exists.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict with the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_TOO_LARGE": {
                    "summary": "PAYLOAD_TOO_LARGE",
                    "value": {
                      "error": {
                        "code": "PAYLOAD_TOO_LARGE",
                        "message": "The request body is too large.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "examples": {
                  "UNSUPPORTED_MEDIA_TYPE": {
                    "summary": "UNSUPPORTED_MEDIA_TYPE",
                    "value": {
                      "error": {
                        "code": "UNSUPPORTED_MEDIA_TYPE",
                        "message": "The content type must be application/json.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "details": [
                          {
                            "field": "title",
                            "message": "Must be between 1 and 120 characters."
                          }
                        ],
                        "message": "One or more fields are invalid.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "RATE_LIMITED": {
                    "summary": "RATE_LIMITED",
                    "value": {
                      "error": {
                        "code": "RATE_LIMITED",
                        "message": "Too many attempts. Try again later.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "summary": "INTERNAL_ERROR",
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "Something went wrong on our side.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "summary": "Register a user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
...
}